+++ src/interfaces/Projectile.java
@@ -3,12 +3,12 @@ package interfaces;
 
 public interface Projectile extends MovingViewInterface{
 
-	public void heal(double howMuchHealed);
-	public double getMaxHealth();
-	public double getHealth();
+	public void heal(int howMuchHealed);
+	public int getMaxHealth();
+	public int getHealth();
 
-	public boolean takeDamage(double amountOfDamage);
-	public void setDamage(double newDamage);
-	public double getDamage();
+	public boolean takeDamage(int amountOfDamage);
+	public void setDamage(int newDamage);
+	public int getDamage();
 	
 }
